+++ src/Spackle.Tests/Extensions/RandomNumberGeneratorTests.cs
@@ -1,21 +1,19 @@
 ﻿using NUnit.Framework;
+using Spackle.Extensions;
 using System.Numerics;
 using System.Security.Cryptography;
 
-namespace Spackle.Tests;
+namespace Spackle.Tests.Extensions;
 
-internal static class SecureRandomTests
+internal static class RandomNumberGeneratorTests
 {
-	[Test]
-	public static void GetBigIntegerWithZeroLength()
-	{
-		var random = new SecureRandom();
-		Assert.That(() => random.GetBigInteger(0),
-			Throws.TypeOf<ArgumentException>()
-				.And.Message.EqualTo("The number of digits cannot be zero. (Parameter 'numberOfDigits')"));
-	}
+   [Test]
+   public static void GetBigIntegerWithZeroLength() => 
+		Assert.That(() => RandomNumberGenerator.Create().GetBigInteger(0),
+		   Throws.TypeOf<ArgumentException>()
+			   .And.Message.EqualTo("The number of digits cannot be zero. (Parameter 'numberOfDigits')"));
 
-	[TestCase(1ul)]
+   [TestCase(1ul)]
 	[TestCase(2ul)]
 	[TestCase(9ul)]
 	[TestCase(10ul)]
@@ -25,7 +23,7 @@ internal static class SecureRandomTests
 	[TestCase(100ul)]
 	public static void GetBigIntegerWithSpecifiedLength(ulong length)
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		var value = random.GetBigInteger(length);
 		Assert.That(value.ToString(), Has.Length.EqualTo((int)length));
 	}
@@ -33,7 +31,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetBigIntegerWithRange()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		var max = BigInteger.Parse("431531631631431");
 		var value = random.GetBigIntegerWithRange(max);
 
@@ -43,7 +41,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetBigIntegerWithRangeWithZero()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetBigIntegerWithRange(BigInteger.Zero),
 			Throws.TypeOf<ArgumentException>()
 				.And.Message.EqualTo("Max value, 0, must be greater than zero. (Parameter 'max')"));
@@ -52,7 +50,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetBigIntegerWithRangeWithNegativeValue()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetBigIntegerWithRange(-3),
 			Throws.TypeOf<ArgumentException>()
 				.And.Message.EqualTo("Max value, -3, must be greater than zero. (Parameter 'max')"));
@@ -63,7 +61,7 @@ internal static class SecureRandomTests
 	{
 		var min = new BigInteger(4500);
 		var max = new BigInteger(5000);
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		var value = random.GetBigIntegerWithRange(min, max);
 
 		Assert.That(value >= min && value < max, Is.True);
@@ -72,7 +70,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetBigIntegerWithMinAndMaxValuesWithMinGreaterThanMax()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetBigIntegerWithRange(4, 3),
 			Throws.TypeOf<ArgumentException>()
 				.And.Message.Contains("Min value, 4, must be less than the max value, 3. (Parameter 'min')"));
@@ -81,7 +79,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetBigIntegerWithMinAndMaxValuesWithMaxLessThanOne()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetBigIntegerWithRange(4, 0),
 			Throws.TypeOf<ArgumentException>()
 				.And.Message.Contains("Max value, 0, must be greater than zero. (Parameter 'max')"));
@@ -90,42 +88,23 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetBigIntegerWithMinAndMaxValuesWithMinLessThanOne()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetBigIntegerWithRange(0, 4),
 			Throws.TypeOf<ArgumentException>()
 				.And.Message.Contains("Min value, 0, must be greater than zero. (Parameter 'min')"));
 	}
 
 	[Test]
-	public static void CreateGeneratorWithDefaultProvider()
-	{
-		var random = new SecureRandom();
-		Assert.That(random.Generator, Is.Not.Null);
-	}
-
-	[Test]
-	public static void CreateGeneratorWithGivenProvider()
-	{
-		using var generator = new MyGenerator();
-		var random = new SecureRandom(generator);
-		Assert.That(random.Generator, Is.Not.Null);
-	}
-
-	[Test]
-	public static void CreateGeneratorWithNullProvider() =>
-		Assert.That(() => new SecureRandom(null!), Throws.TypeOf<ArgumentNullException>());
-
-	[Test]
 	public static void GenerateIntegerWithNegativeUpperBound()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.Next(-2), Throws.TypeOf<ArgumentException>());
 	}
 
 	[Test]
 	public static void GenerateIntegers()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		for (var i = 0; i < 100000; i++)
 		{
 			var x = random.Next();
@@ -136,7 +115,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegersWithMaxLimit()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		const int Max = 25;
 
 		for (var i = 0; i < 100000; i++)
@@ -153,7 +132,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegerWithNegativeMaxValue()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.Next(-1),
 			Throws.TypeOf<ArgumentException>()
 				.And.Message.EqualTo("maxValue must be greater than or equal to zero. (Parameter 'maxValue')"));
@@ -162,7 +141,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegersWithPositiveMaxAndMinLimits()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		const int Max = 25;
 		const int Min = 15;
 
@@ -180,7 +159,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegersWithPositiveMaxAndNegativeMinLimits()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		const int Max = 25;
 		const int Min = -15;
 
@@ -198,7 +177,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegersWithNegativeMaxAndMinLimits()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		const int Max = -15;
 		const int Min = -25;
 
@@ -216,7 +195,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegersWithSameMaxAndMinLimits()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		const int Max = 25;
 		const int Min = 25;
 
@@ -229,7 +208,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateIntegersWithInvalidMaxAndMinValues()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		const int Max = -25;
 		const int Min = 25;
 		Assert.That(() => random.Next(Min, Max),
@@ -240,14 +219,14 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GenerateBits()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		random.NextBytes(new byte[10]);
 	}
 
 	[Test]
 	public static void GenerateBitsWithNullArgument()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.NextBytes(null!), Throws.TypeOf<ArgumentNullException>());
 	}
 
@@ -256,14 +235,13 @@ internal static class SecureRandomTests
 	public static void GenerateBoolean(bool value)
 	{
 		using var generator = new MockBoolGenerator(value);
-		var random = new SecureRandom(generator);
-		Assert.That(random.NextBoolean(), Is.EqualTo(value));
+		Assert.That(generator.NextBoolean(), Is.EqualTo(value));
 	}
 
 	[Test]
 	public static void GenerateDoubles()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		for (var i = 0; i < 500000; i++)
 		{
 			var d = random.NextDouble();
@@ -279,9 +257,8 @@ internal static class SecureRandomTests
 	public static void GetByteValuesDuplicatesAllowed()
 	{
 		using var generator = new MockRandomNumberGeneratorForGetByteValues(ValueGeneration.DuplicatesAllowed);
-		var random = new SecureRandom(generator);
 
-		var elements = random.GetByteValues(10, ValueGeneration.DuplicatesAllowed);
+		var elements = generator.GetByteValues(10, ValueGeneration.DuplicatesAllowed);
 		{
 			using (Assert.EnterMultipleScope())
 			{
@@ -302,7 +279,7 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetByteValuesDuplicatesAllowedAndElementNumberExceedsByteMaximum()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		var elements = random.GetByteValues(2560, ValueGeneration.DuplicatesAllowed);
 		using (Assert.EnterMultipleScope())
 		{
@@ -323,8 +300,7 @@ internal static class SecureRandomTests
 	public static void GetByteValuesUniqueValuesOnly()
 	{
 		using var generator = new MockRandomNumberGeneratorForGetByteValues(ValueGeneration.UniqueValuesOnly);
-		var random = new SecureRandom(generator);
-		var elements = random.GetByteValues(10, ValueGeneration.UniqueValuesOnly);
+		var elements = generator.GetByteValues(10, ValueGeneration.UniqueValuesOnly);
 		using (Assert.EnterMultipleScope())
 		{
 			Assert.That(elements, Has.Length.EqualTo(10), nameof(elements.Length));
@@ -344,14 +320,14 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetByteValuesUniqueValuesOnlyButElementNumberIsTooBig()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetByteValues(2560, ValueGeneration.UniqueValuesOnly), Throws.TypeOf<ArgumentException>());
 	}
 
 	[Test]
 	public static void GetDoubleValues()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		var elements = random.GetDoubleValues(8);
 		using (Assert.EnterMultipleScope())
 		{
@@ -368,8 +344,7 @@ internal static class SecureRandomTests
 	public static void GetInt32ValuesDuplicatesAllowed()
 	{
 		using var generator = new MockRandomNumberGeneratorForGetInt32Values(ValueGeneration.DuplicatesAllowed);
-		var random = new SecureRandom(generator);
-		var elements = random.GetInt32Values(8, ValueGeneration.DuplicatesAllowed);
+		var elements = generator.GetInt32Values(8, ValueGeneration.DuplicatesAllowed);
 		using (Assert.EnterMultipleScope())
 		{
 			Assert.That(elements, Has.Length.EqualTo(8), nameof(elements.Length));
@@ -389,8 +364,7 @@ internal static class SecureRandomTests
 	public static void GetInt32ValuesUniqueValuesOnly()
 	{
 		using var generator = new MockRandomNumberGeneratorForGetInt32Values(ValueGeneration.UniqueValuesOnly);
-		var random = new SecureRandom(generator);
-		var elements = random.GetInt32Values(8, ValueGeneration.UniqueValuesOnly);
+		var elements = generator.GetInt32Values(8, ValueGeneration.UniqueValuesOnly);
 		using (Assert.EnterMultipleScope())
 		{
 			Assert.That(elements, Has.Length.EqualTo(8), nameof(elements.Length));
@@ -410,22 +384,9 @@ internal static class SecureRandomTests
 	[Test]
 	public static void GetInt32ValuesUniqueValuesOnlyButElementNumberIsTooBig()
 	{
-		var random = new SecureRandom();
+		using var random = RandomNumberGenerator.Create();
 		Assert.That(() => random.GetInt32Values(int.MaxValue + 10u, ValueGeneration.UniqueValuesOnly),
 			Throws.TypeOf<ArgumentException>());
-	}
-
-	private sealed class MyGenerator
-		: RandomNumberGenerator
-	{
-#pragma warning disable CA2215 // Dispose methods should call base class dispose
-		protected override void Dispose(bool disposing) => this.DisposeCount++;
-#pragma warning restore CA2215 // Dispose methods should call base class dispose
-
-		public override void GetBytes(byte[] data) =>
-			 throw new NotImplementedException();
-
-		public int DisposeCount { get; private set; }
 	}
 
 	private sealed class MockBoolGenerator